requirements.txt
langchain
langchain-community
langchain-experimental
fastapi
python-dotenv 
streamlit 
uvicorn 
pydantic 
httpx
requests 
langchain_google_community 
langchain_tavily
langchain_openai 
langgraph
langchain-groq
langchain-google-community[places]
-e . 